+++ vue/src/components/Dialogs/Setup/Deployments/UpdateDialogs/DeploymentUpdateImagePullPolicyDialog.vue
@@ -1,0 +1,138 @@
+<script setup lang="ts">
+import {computed, defineComponent, onMounted, onUnmounted, reactive, ref, watch} from 'vue'
+import {Deployment} from "@/core/services/Deploy/models";
+import {Api} from "@/core/services/Deploy/Api";
+import bus from "@/plugins/bus";
+import type {DialogEventsInterface} from "@/components/Dialogs/DialogEventsInterface";
+import {ImagePullPolicies} from "@/constants";
+
+export interface DeploymentUpdateImagePullPolicyDialog_Input {
+    deployment: Deployment;
+}
+
+const props = defineProps<{ input: DeploymentUpdateImagePullPolicyDialog_Input, events: DialogEventsInterface }>();
+
+const used = ref(false);
+const showDialog = ref(false);
+
+const value = ref<string>();
+const imagePullPolicies = ref([
+    {
+        identifier: ImagePullPolicies.IfNotPresent,
+        name: "If not present",
+    },
+    {
+        identifier: ImagePullPolicies.Always,
+        name: "Always",
+    },
+    {
+        identifier: ImagePullPolicies.Never,
+        name: "Never",
+    },
+]);
+
+// <editor-fold desc="Functions">
+
+onMounted(() => {
+    if (used.value) {
+        return;
+    }
+    used.value = true;
+    render();
+});
+
+onUnmounted(() => {
+});
+
+function render() {
+    value.value = props.input.deployment.image_pull_policy ?? '';
+    showDialog.value = true;
+}
+
+function close() {
+    showDialog.value = false;
+    props.events.onClose();
+}
+
+// </editor-fold>
+
+// <editor-fold desc="View Binding Functions">
+
+function onSaveBtnClicked() {
+    const api = Api.deployments().updateImagePullPolicyPutById(props.input.deployment.id!)
+        .value(value.value!)
+    api.setErrorHandler(response => {
+        if (response.error) {
+            bus.emit('toast', {
+                text: response.error
+            });
+        }
+        return false;
+    });
+    api.save(null, newItem => {
+        bus.emit('deploymentSaved', newItem);
+        close();
+    });
+
+    close();
+}
+
+function onCloseBtnClicked() {
+    close();
+}
+
+// </editor-fold>
+
+</script>
+
+<template>
+    <v-dialog
+        persistent
+        width="60vw"
+        v-model="showDialog">
+        <v-card
+            class="w-100 h-100">
+            <v-card-title>Deployment</v-card-title>
+            <v-divider/>
+            <v-card-text>
+                <v-row
+                    dense>
+                    <v-col cols="12">
+                        <v-select
+                            v-model="value"
+                            :items="imagePullPolicies"
+                            item-title="name"
+                            item-value="identifier"
+                            variant="outlined"
+                            label="Image Pull Policy"
+                        />
+                    </v-col>
+                </v-row>
+            </v-card-text>
+            <v-divider/>
+            <v-card-actions>
+                <v-spacer/>
+                <v-btn
+                    variant="tonal"
+                    color="grey"
+                    prepend-icon="fa fa-circle-xmark"
+                    @click="onCloseBtnClicked">
+                    Close
+                </v-btn>
+
+                <v-btn
+                    flat
+                    variant="tonal"
+                    prepend-icon="fa fa-check"
+                    color="green"
+                    @click="onSaveBtnClicked">
+                    Save
+                </v-btn>
+            </v-card-actions>
+        </v-card>
+    </v-dialog>
+</template>
+
+<style scoped>
+
+</style>