+++ vue/src/core/services/Deploy/models/definitions/ContainerImageDefinition.ts
@@ -10,6 +10,7 @@ export class ContainerImageDefinition extends BaseModel {
     url?: string;
     pull_secret?: string;
     default_tag?: string;
+    default_image_pull_policy?: string;
     registry_subscribe?: boolean;
     registry_provider?: string;
     registry_provider_gcloud_registry_name?: string;
@@ -54,6 +55,7 @@ export class ContainerImageDefinition extends BaseModel {
             delete this.url;
             delete this.pull_secret;
             delete this.default_tag;
+            delete this.default_image_pull_policy;
             delete this.registry_subscribe;
             delete this.registry_provider;
             delete this.registry_provider_gcloud_registry_name;
@@ -100,6 +102,9 @@ export class ContainerImageDefinition extends BaseModel {
         }
         if (data.default_tag != null) {
             this.default_tag = data.default_tag;
+        }
+        if (data.default_image_pull_policy != null) {
+            this.default_image_pull_policy = data.default_image_pull_policy;
         }
         if (data.registry_subscribe != null) {
             this.registry_subscribe = data.registry_subscribe;